# backend/tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "AppointmentCancelRequest": {
        "properties": {
          "cancelled_reason": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Cancelled Reason",
            "type": "string"
          }
        },
        "required": [
          "cancelled_reason"
        ],
        "title": "AppointmentCancelRequest",
        "type": "object"
      },
      "AppointmentCreateRequest": {
        "properties": {
          "appointment_date": {
            "format": "date",
            "title": "Appointment Date",
            "type": "string"
          },
          "appointment_type": {
            "enum": [
              "new",
              "follow_up",
              "emergency"
            ],
            "title": "Appointment Type",
            "type": "string"
          },
          "doctor_id": {
            "format": "uuid",
            "title": "Doctor Id",
            "type": "string"
          },
          "end_time": {
            "format": "time",
            "title": "End Time",
            "type": "string"
          },
          "is_walk_in": {
            "default": false,
            "title": "Is Walk In",
            "type": "boolean"
          },
          "location_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "patient_id": {
            "format": "uuid",
            "title": "Patient Id",
            "type": "string"
          },
          "start_time": {
            "format": "time",
            "title": "Start Time",
            "type": "string"
          }
        },
        "required": [
          "patient_id",
          "doctor_id",
          "appointment_date",
          "start_time",
          "end_time",
          "appointment_type"
        ],
        "title": "AppointmentCreateRequest",
        "type": "object"
      },
      "AppointmentUpdateRequest": {
        "properties": {
          "appointment_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appointment Date"
          },
          "appointment_type": {
            "anyOf": [
              {
                "enum": [
                  "new",
                  "follow_up",
                  "emergency"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appointment Type"
          },
          "end_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "location_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "start_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          }
        },
        "title": "AppointmentUpdateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Hospital Management SaaS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/appointments": {
      "get": {
        "description": "List appointments with optional filters \u2014 requires appointment:read.",
        "operationId": "list_appointments_api_v1_appointments_get",
        "parameters": [
          {
            "in": "query",
            "name": "doctor_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Doctor Id"
            }
          },
          {
            "in": "query",
            "name": "patient_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Id"
            }
          },
          {
            "in": "query",
            "name": "appointment_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Appointment Date"
            }
          },
          {
            "in": "query",
            "name": "from_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Date"
            }
          },
          {
            "in": "query",
            "name": "to_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Date"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "description": "Page number (1-based)",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number (1-based)",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page",
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Appointments Api V1 Appointments Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Appointments",
        "tags": [
          "appointments"
        ]
      },
      "post": {
        "description": "Book a new appointment \u2014 requires appointment:create.",
        "operationId": "create_appointment_api_v1_appointments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Appointment",
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/v1/appointments/availability": {
      "get": {
        "description": "Return bookable slots for a doctor on a date \u2014 requires appointment:read.",
        "operationId": "get_appointment_availability_api_v1_appointments_availability_get",
        "parameters": [
          {
            "description": "Doctor to check availability for",
            "in": "query",
            "name": "doctor_id",
            "required": true,
            "schema": {
              "description": "Doctor to check availability for",
              "format": "uuid",
              "title": "Doctor Id",
              "type": "string"
            }
          },
          {
            "description": "Date to check (ISO)",
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "description": "Date to check (ISO)",
              "format": "date",
              "title": "Date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "location_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Appointment Availability Api V1 Appointments Availability Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Appointment Availability",
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/v1/appointments/{appointment_id}": {
      "delete": {
        "description": "Soft-delete an appointment \u2014 requires appointment:update.",
        "operationId": "delete_appointment_api_v1_appointments__appointment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Appointment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Appointment",
        "tags": [
          "appointments"
        ]
      },
      "get": {
        "description": "Get appointment details \u2014 requires appointment:read.",
        "operationId": "get_appointment_api_v1_appointments__appointment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Appointment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Appointment Api V1 Appointments  Appointment Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Appointment",
        "tags": [
          "appointments"
        ]
      },
      "patch": {
        "description": "Reschedule or update an appointment \u2014 requires appointment:update.",
        "operationId": "update_appointment_api_v1_appointments__appointment_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Appointment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Appointment Api V1 Appointments  Appointment Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Appointment",
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/v1/appointments/{appointment_id}/cancel": {
      "post": {
        "description": "Cancel an appointment with a reason \u2014 requires appointment:update.",
        "operationId": "cancel_appointment_api_v1_appointments__appointment_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Appointment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cancel Appointment Api V1 Appointments  Appointment Id  Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Appointment",
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/v1/appointments/{appointment_id}/confirm": {
      "post": {
        "description": "Confirm a scheduled appointment \u2014 requires appointment:update.",
        "operationId": "confirm_appointment_api_v1_appointments__appointment_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Appointment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Confirm Appointment Api V1 Appointments  Appointment Id  Confirm Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Confirm Appointment",
        "tags": [
          "appointments"
        ]
      }
    },
...
  }
}